requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.5
filelock==3.12.4
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.26.0
opencv-python==4.8.1.78
Pillow==10.0.1
requests==2.31.0
sqlparse==0.4.4
sympy==1.12
torch==2.0.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6